Add an option to rerun a query to eclcc regression script

+++ ecl/regress/regress.sh
@@ -27,7 +27,7 @@
 # in your build directory. (option: -e $BUILDDIR/Debug/bin/eclcc)
 ##############################################################################
 
-syntax="syntax: $0 [-t target_dir] [-c compare_dir] [-I include_dir ...] [-e eclcc] [-d diff_program]"
+syntax="syntax: $0 [-t target_dir] [-c compare_dir] [-I include_dir ...] [-e eclcc] [-d diff_program] [-q query.ecl]"
 
 ## Default arguments
 target_dir=run_$$
@@ -37,6 +37,7 @@ compare_only=0
 userflags=
 eclcc=
 diff=
+query=
 np=`grep -c processor /proc/cpuinfo`
 export ECLCC_ECLINCLUDE_PATH=
 
@@ -49,11 +50,12 @@ if [[ $1 = '' ]]; then
     echo " * include dir for special ECL headers (allows multiple paths)"
     echo " * eclcc necessary for compilation, otherwise, only comparison will be made"
     echo " * diff_program must be able to handle directories"
+    echo " * -q can be used to run/rerun a single query"
     echo
     exit -1
 fi
 if [[ $* != '' ]]; then
-    while getopts "t:c:I:e:d:f:" opt; do
+    while getopts "t:c:I:e:d:f:q:" opt; do
         case $opt in
             t)
                 target_dir=$OPTARG
@@ -73,6 +75,9 @@ if [[ $* != '' ]]; then
             f)
                 userflags="$userflags -f$OPTARG"
                 ;;
+            q)
+                query="$OPTARG"
+                ;;
             :)
                 echo $syntax
                 exit -1
@@ -87,7 +92,9 @@ if [[ $eclcc != '' ]]; then
     flags="$default_flags $include_dir -fshowMetaInGraph $userflags"
 
     ## Prepare target directory
-    rm -rf $target_dir
+    if [[ $query == '' ]]; then
+        rm -rf $target_dir
+    fi
     mkdir -p $target_dir
 
     ## Create Makefile (git doesn't like tabs)
@@ -110,11 +117,17 @@ if [[ $eclcc != '' ]]; then
     echo "\$(TARGET)/%.eclxml.log: %.eclxml" >> Makefile
     echo -e "\t\$(ECLCC) \$(FLAGS) $^" >> Makefile
 
-    ## Compile all regressions
-    echo "* Compiling all regression tests"
-    echo
-    export eclcc flags include_dir target_dir
-    time make -j $np > /dev/null
+    if [[ $query != '' ]]; then
+        ## Compile all regressions
+        echo "Compiling $query regression test"
+        echo
+        time make $query.run -B > /dev/null
+    else
+        ## Compile all regressions
+        echo "* Compiling all regression tests"
+        echo
+        time make -j $np > /dev/null
+    fi
 fi
 
 ## Compare to golden standard (ignore obvious differences)
